requirements.txt
streamlit
altair
faiss-cpu
sentence-transformers
matplotlib
seaborn>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
statsmodels>=0.12.0
thefuzz>=0.19.0
polyfuzz>=0.4.2
scikit-learn>=0.24.0
joblib>=1.0.0
tqdm>=4.41.1
jinja2>=3.0
networkx>=2.5
pysam>=0.16.0
sqlalchemy>=1.4
cython>=0.29
PyYAML>=5.0